+++ bank-feeds/src/codat_bankfeeds/models/source_account_v2.py
@@ -23,11 +23,11 @@ import re  # noqa: F401
 import json
 
 from pydantic import BeforeValidator, PlainSerializer, BaseModel, ConfigDict, Field, field_validator, model_serializer
-from typing import Annotated, Any, ClassVar, Dict, List, Optional, Union
+from typing import Any, ClassVar, Dict, List, Optional, Union
 from codat_bankfeeds.models.shared.accountinfo import AccountInfo, AccountInfoTypedDict
 from codat_bankfeeds.models.shared.routinginfo import RoutingInfo, RoutingInfoTypedDict
 from typing import Optional, Set
-from typing_extensions import Self, NotRequired, TypedDict
+from typing_extensions import Annotated, Self, NotRequired, TypedDict
 
 from codat_bankfeeds.models._compat.sourceaccountv2 import SourceAccountV2AccountType
 from codat_bankfeeds.types import OptionalNullable, UNSET, UNSET_SENTINEL